se agrego multer aun que al final no se uso, tambien se eliminaron las validacion que se hacian con express-validator y ahora se usa class-validator desde dominio y application, se realizaron las pruebas y todos los enpoints son funcionales, solo falta agregar desactivar status

+++ src/user/application/deleteUserUseCase.ts
@@ -1,11 +1,18 @@
+import { validate } from "class-validator";
 import { IUsuarioRepository } from "../domain/userRepository";
+import { ValidatorId } from "../domain/validations/user";
 
 
 export class DeleteUserUseCase {
     constructor(readonly usuarioRepository: IUsuarioRepository) {}
     
     async run(uuid: string,): Promise<string | null> {
-      
+        let post = new ValidatorId(uuid)
+        const validation = await validate(post)
+        console.log(validation.length)
+        if (validation.length > 0) {
+            throw new Error(JSON.stringify(validation));
+        }
         try {
             const deleteUser = await this.usuarioRepository.deleteUserById(uuid);
             return deleteUser;
